requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
async-property==0.2.2
attrs==25.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-storage==22.0.0
azure-mgmt-web==7.3.1
azure-storage-blob==12.23.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.1
cryptography==43.0.1
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.0
fastapi-cli==0.0.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
ipykernel==6.29.2
ipython==8.22.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jwcrypto==1.5.6
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
nest-asyncio==1.6.0
nodeenv==1.9.1
oauthlib==3.2.2
orjson==3.10.7
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
portalocker==2.10.1
prisma==0.15.0
prompt-toolkit==3.0.43
propcache==0.2.1
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-keycloak==4.6.2
python-multipart==0.0.12
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
sendgrid==6.11.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.38.6
tenacity==9.0.0
tomlkit==0.13.2
tornado==6.4
traitlets==5.14.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.1
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.18.3
